Adds initial tests of using router.js.
At this point I'm not so sure how to get app.js and router.js to work
together. One of the resources  I'm using is this conversatin on SO:

https://stackoverflow.com/questions/28143419/express-static-keeps-routing-my-
files-from-the-route

I am also looking at:

https://www.freecodecamp.org/news/how-to-write-a-production-ready-node-
and-express-app-f214f0b17d8c/

For now, I'm going to focus on working within app.js, and then refactor
the code to have app.js and router.js work together properly.